requirements.txt
pillow
requests
lxml
